+++ CHANGELOG.md
@@ -1,9 +1,17 @@
+1.8.0 / 2019-01-30
+==================
+
+  * FindCssPath and findXPath
+  * Delay in type
+  * Browser.requests.getAllMocks method
+  * Several dependencies updated
+
 1.7.2 / 2019-01-12
 ==================
 
   * WaitUntilEnabled
   * Changed browser.assert.url error message when internal error happens
-  * Fixed unexpected behaviour on closing a browser that has never been opened
+  * Fixed unexpected behavior on closing a browser that has never been opened
 
 1.7.1 / 2019-01-08
 ==================